requirements.txt
python-dotenv
google-generativeai
langchain
langchain_community
langchain_openai
langchain_google_genai
PyPDF2
faiss-cpu
streamlit
-e .